+++ bwapi/Frontend/BWAPILIBUnitTest/UnitFinderFixture.h
@@ -19,7 +19,7 @@ public:
 protected:
   BWAPI::UnitFinder finder;
 
-  int countUnits(int left = 0, int top = 0, int right = 10000, int bottom = 10000)
+  size_t countUnits(int left = 0, int top = 0, int right = 10000, int bottom = 10000)
   {
     return finder.find(left, top, right, bottom).size();
   }
